requirements.txt
absl-py==2.3.1
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cryptography==41.0.7
cycler==0.12.1
dnspython==2.7.0
filelock==3.18.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
flask-mysql-connector==1.1.0
Flask-MySQLdb==2.0.0
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.0.5
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.7.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.76.0
h5py==3.15.1
huggingface-hub==0.34.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
mlxtend==0.23.4
mongoengine==0.29.1
mpmath==1.3.0
mysql-connector==2.2.9
mysql-connector-python==9.3.0
mysqlclient==2.2.7
namex==0.1.0
narwhals==2.6.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
patsy==1.0.1
pillow==11.3.0
playsound==1.3.0
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.13.2
PyMySQL==1.1.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rich==14.2.0
rpds-py==0.27.1
safetensors==0.6.1
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.42
statsmodels==0.14.5
streamlit==1.50.0
summa==1.2.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
tokenizers==0.21.4
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.55.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
waitress==3.0.2
watchdog==6.0.0
Werkzeug==2.3.7
wheel==0.45.1
wrapt==2.0.0
